Add wheel slip plugin to vehicles (#569)
* Add wheel slip plugin to vehicles

The Wheel slip plugin dynamically adjusts the slip parameter of wheels
in proportion to the speed of the vehicle. This is a better (i.e. more
realistic) model of tires than the raw slip parameter in <surface>
elements. When the vehicle is not moving, the plugin sets the slip
parameter to 0, which prevents sliding.

I have set the `<slip_compliance_lateral>` lateral parameter
experimentally to maintain the current turning behavior of the vehicles.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>

* Added plugin to other launch files

Signed-off-by: Nate Koenig <nate@openrobotics.org>

* Add wheel slip plugin to ctu_cras and updates to other models to improve steering

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>

* More improvements of steering in place for Husky robots

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>

Co-authored-by: Nate Koenig <nate@openrobotics.org>