Add flipper torque limit control (#948)
Implements the so-far-missing support for torque limit control for robots with flippers. Having this allows making much better controllers.Having this allows making much better controllers.

The code explicitly checks if the torque limit isn't larger than the `<joint effort="">` attribute in SDF, so it shouldn't be possible to create unrealistically strong flippers.